refactor and add timed effect
